+++ CHANGELOG.md
@@ -3,7 +3,7 @@ Changelog
 
 New Features
 
-- Uses Babin's ``map2disc`` package to find nested initial coordinate mappings for complex boundaries.
+- Uses Babin's harmonic mapping (via the ``map2disc-jax`` package) to find nested initial coordinate mappings for complex boundaries. Select with ``ensure_nested_method="map2disc"``.
 - Added warning for when ``deriv_mode="batched"`` is used in an ``ObjectiveFunction`` where one or more sub-objectives is using ``rev`` mode differentiation. Also adds more info about the derivative mode and Jacobian chunk sizes when building the objective with ``verbose>1``.
 
 Performance Improvements